Tests and Bug fix
Added:
- MSUnit tests for Faker.Core library
- Additional classes for testing

Bug fix:
- Faker ctor has public modificator
- DecimalGenerator has explicit conversion
- GetDefaultValue is now public (for access in test)

+++ Faker/Faker.Core/TypeGenerators/DecimalGenerator.cs
@@ -11,6 +11,6 @@ public class DecimalGenerator : IValueGenerator
     // Generate random decimal
     public object Generate(Type typeToGenerate, GeneratorContext context)
     {
-        return context.Random.Next(int.MinValue, int.MaxValue);
+        return (decimal)context.Random.Next(int.MinValue, int.MaxValue);
     }
 }
